test: refactored tests and parent-child creation
